allow user to click on column headers to sort endpoints
See https://trello.com/c/GNAHdqpw/15-endpoint-list-in-addition-to-the-order-by-field-is-it-possible-to-let-the-user-click-on-the-column-header-to-sort-eg-study-expe

This includes animal endpoints, invitro endpoints, and epi outcomes. Does not include epi meta. Essentially
this commit just adds some shortcut links to the headers.

Basic conversion approach for each type of outcome:
1. add column names/sortkeys to relevant Django forms.py file, if needed
2. add a hidden field to the form template to store the passed in order_by value
3. add entry to the JS that sets up the table to establish mapping between column names and sort keys

(and as a one time, modify BaseTable.js with some utility functions, add some new css, etc.)

Note that epi outcomes already had sorting. For consistency's sake, that's been replaced with this
new approach which includes things like a sort indicator on the column.