add study short citation to epi/animal/invitro bulk cleaning interfaces
See https://trello.com/c/eB81wYux/17-for-clean-extracted-data-can-you-add-the-study-name-to-the-expanded-fields-it-is-not-possible-right-now-to-determine-what-data

basic approach for each Serializer (e.g. StudyPopulationCleanupFieldsSerializer):
1. add a new SerializerMethodField called "study_short_citation"
2. add this field to "cleanup_fields" / "fields"
3. define a method that grabs this through whatever object chain is needed for the type in question

Preferred approach would be to just access field names ORM style, like "experiment__study__short_citation"
but I could not figure out an approach to make that work.

Also looked into an approach using a mixin that could define the #3 method and just delegate it to an
appropriate method on the model class. But I was unable to get this approach to work either, likely
because I don't fully understand Django's rest_framework inner workings.