implement study level locking
See https://trello.com/c/R6X7KJZD/58-locking-data-at-the-study-level

Adds an "editable" field on the Study class which can be set by users with
assessment-edit permissions via the "Actions" dropdown on the study detail page
(called "Lock Study" / "Unlock Study")

This setting will be considered by HAWC's permission system when deciding if a user
can edit a study or a related item to that study. For example:
- if a user is looking at a locked study or an item related to it like a study population,
  they will not see the "Actions" dropdown.
- if a user is already on a page for a study when it is locked, and then they attempt to
  edit/delete/etc. it, they will get a 403/Forbidden error.
- superusers can still edit anything regardless of lock status

The study name in the breadcrumbs will display a locked icon if the study is locked via this
mechanism. This could be extended to show a lock icon for uneditable assessments as well, if desired.